Refactor Query API, adjust imports and add documentation
This commit includes a cleanup of the query API by moving some classes to specific namespaces, adjusting the imports as necessary. Simultaneously, improvements have been made to code documentation for enhanced readability and understanding of the classes 'SingletonCall.java' and 'BatchCall.java'. The changes contribute to simpler future code interactions and improved maintainability.
