+++ sadu-queries2/src/main/java/de/chojo/sadu/queries/stages/ParsedQueryImpl.java
@@ -6,12 +6,12 @@
 
 package de.chojo.sadu.queries.stages;
 
-import de.chojo.sadu.queries.TokenizedQuery;
-import de.chojo.sadu.queries.api.ParsedQuery;
 import de.chojo.sadu.queries.api.execution.writing.CalledBatchQuery;
 import de.chojo.sadu.queries.api.execution.writing.CalledSingletonQuery;
+import de.chojo.sadu.queries.api.query.ParsedQuery;
 import de.chojo.sadu.queries.calls.BatchCall;
 import de.chojo.sadu.queries.calls.SingletonCall;
+import de.chojo.sadu.queries.query.TokenizedQuery;
 import de.chojo.sadu.queries.stages.base.QueryProvider;
 import de.chojo.sadu.queries.stages.calls.CallSupplier;
 import de.chojo.sadu.queries.stages.execution.writing.CalledBatchQueryImpl;
