+++ sadu-queries2/src/main/java/de/chojo/sadu/queries/stages/calls/CallSupplier.java
@@ -6,7 +6,7 @@
 
 package de.chojo.sadu.queries.stages.calls;
 
-import de.chojo.sadu.queries.calls.Calls;
+import de.chojo.sadu.queries.api.call.calls.Calls;
 import de.chojo.sadu.queries.storage.ResultStorage;
 
 @FunctionalInterface
